オートモードhooks化
